Organises Extensions in their correct files
