+++ Supermarket/Extensions/Bool+Extensions.swift
