+++ Supermarket/Extensions/Color+Extensions.swift
@@ -1,0 +1,19 @@
+//
+//  Color+Extensions.swift
+//  Supermarket
+//
+//  Created by Mohcine on 19/11/2024.
+//
+
+import SwiftUI
+
+extension Color {
+    static var background: Color {
+        .init(red: 239.0/255, green: 239.0/255, blue: 244.0/255)
+    }
+    
+    static var darkGreen: Color {
+        .init(red: 0, green: 85.0/255, blue: 20.0/255)
+    }
+    
+}
