+++ Supermarket/Extensions/Double+Extensions.swift
